+++ pyaud_plugins/__init__.py
@@ -3,6 +3,7 @@ pyaud_plugins
 =============
 """
 from . import modules
+from ._environ import environ
 from ._version import __version__
 
-__all__ = ["__version__", "modules"]
+__all__ = ["__version__", "modules", "environ"]