+++ tests/_test.py
@@ -80,7 +80,9 @@ def test_write_command(
 
         def mock_write_whitelist(*_: t.Any, **__: t.Any) -> None:
             with open(
-                Path.cwd() / pyaud.environ.WHITELIST, "w", encoding="utf-8"
+                Path.cwd() / pyaud_plugins.environ.WHITELIST,
+                "w",
+                encoding="utf-8",
             ) as fout:
                 fout.write(content)
 
@@ -246,7 +248,7 @@ def test_make_docs_rm_cache(
     :param call_status: Patch function to return specific exit-code.
     :param make_tree: Create directory tree from dict mapping.
     """
-    builddir = Path.cwd() / pyaud.environ.BUILDDIR
+    builddir = Path.cwd() / pyaud_plugins.environ.BUILDDIR
     readme = Path.cwd() / README
 
     # disable call to ``Subprocess`` to only create ./docs/_build
@@ -330,7 +332,7 @@ def test_pipfile2req_commands(
     :param nocolorcapsys: Capture system output while stripping ANSI
         color codes.
     """
-    requirements = Path.cwd() / pyaud.environ.REQUIREMENTS
+    requirements = Path.cwd() / pyaud_plugins.environ.REQUIREMENTS
     pipfile_lock = Path.cwd() / PIPFILE_LOCK
     with open(pipfile_lock, "w", encoding="utf-8") as fout:
         fout.write(files.PIPFILE_LOCK)
@@ -388,7 +390,7 @@ def test_append_whitelist(
         announce what is called.
     """
     project_dir = Path.cwd()
-    whitelist = project_dir / pyaud.environ.WHITELIST
+    whitelist = project_dir / pyaud_plugins.environ.WHITELIST
     Path(project_dir / FILES).touch()
     whitelist.touch()
     pyaud.git.add(".")
@@ -521,7 +523,7 @@ def test_make_requirements(
     :param nocolorcapsys: Capture system output while stripping ANSI
         color codes.
     """
-    path = Path.cwd() / pyaud.environ.REQUIREMENTS
+    path = Path.cwd() / pyaud_plugins.environ.REQUIREMENTS
     with open(Path.cwd() / PIPFILE_LOCK, "w", encoding="utf-8") as fout:
         fout.write(files.PIPFILE_LOCK)
 
@@ -550,7 +552,7 @@ def test_make_whitelist(
     :param make_tree: Create directory tree from dict mapping.
     """
     project_dir = Path.cwd()
-    whitelist = project_dir / pyaud.environ.WHITELIST
+    whitelist = project_dir / pyaud_plugins.environ.WHITELIST
     make_tree(
         project_dir,
         {
@@ -750,7 +752,9 @@ def test_deploy_master(
     mock_plugins = pyaud.plugins.mapping()
 
     def _docs(*_: t.Any, **__: t.Any):
-        Path(Path.cwd() / pyaud.environ.BUILDDIR / "html").mkdir(parents=True)
+        Path(Path.cwd() / pyaud_plugins.environ.BUILDDIR / "html").mkdir(
+            parents=True
+        )
 
     mock_plugins["docs"] = _docs  # type: ignore
     monkeypatch.setattr(PYAUD_PLUGINS_PLUGINS, mock_plugins)
@@ -819,7 +823,9 @@ def test_deploy_master_param(
     mock_plugins = pyaud.plugins.mapping()
 
     def _docs(*_: t.Any, **__: t.Any) -> None:
-        Path(path / pyaud.environ.BUILDDIR / "html").mkdir(parents=True)
+        Path(path / pyaud_plugins.environ.BUILDDIR / "html").mkdir(
+            parents=True
+        )
 
     mock_plugins["docs"] = _docs  # type: ignore
     monkeypatch.setattr(PYAUD_PLUGINS_PLUGINS, mock_plugins)
@@ -854,7 +860,7 @@ def test_deploy_cov_report_token(
     :param patch_sp_print_called: Patch ``Subprocess.call`` to only
         announce what is called.
     """
-    Path(Path.cwd() / pyaud.environ.COVERAGE_XML).touch()
+    Path(Path.cwd() / pyaud_plugins.environ.COVERAGE_XML).touch()
     patch_sp_print_called()
     monkeypatch.setenv("CODECOV_TOKEN", "token")
     main("deploy-cov")
@@ -874,7 +880,7 @@ def test_deploy_cov_no_token(
     :param nocolorcapsys: Capture system output while stripping ANSI
         color codes.
     """
-    Path(Path.cwd() / pyaud.environ.COVERAGE_XML).touch()
+    Path(Path.cwd() / pyaud_plugins.environ.COVERAGE_XML).touch()
     main("deploy-cov")
     out = nocolorcapsys.stdout()
     assert "CODECOV_TOKEN not set" in out
@@ -1028,10 +1034,12 @@ def test_make_unused_fix(
         "Updating ``{}``\n"
         "created ``whitelist.py``\n"
         "Success: no issues found in 1 source files\n".format(
-            file, Path.cwd() / pyaud.environ.WHITELIST
+            file, Path.cwd() / pyaud_plugins.environ.WHITELIST
         )
     )
-    with open(Path.cwd() / pyaud.environ.WHITELIST, encoding="utf-8") as fin:
+    with open(
+        Path.cwd() / pyaud_plugins.environ.WHITELIST, encoding="utf-8"
+    ) as fin:
         assert fin.read().strip() == (
             "reformat_this  # unused function (repo/file.py:1)"
         )