+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.1.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Adds `pyaud_plugins.environ`
 
 [0.1.0](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.1.0) - 2022-01-09
 ------------------------------------------------------------------------